Added the window.confirm.
For some reason the delete button does not show up before refreshing the page even though the user who created the post is logged in.

+++ Osa5/bloglist-frontend/src/components/Blog.js
@@ -26,7 +26,9 @@ const Blog = ({blog, handleAddLike, handleRemove, username}) => {
   }
 
   const deleteBlog = () => {
-    handleRemove(blog.id)
+    if (window.confirm(`Want to delete ${blog.author}: ${blog.title}?`)) {
+      handleRemove(blog.id)
+    }
   }
 
   const deleteButton = () => {
